Modified result calculation methodology
Previously, the reliability of calculated score was unknown.  So changed
the result calculation methodology as follow so that we can get
confidence intervals.  The modified script analysis.py returns 95%
confidence intervals as well as scores for each test samples.

A score is calculated as a rate of wins.
When a sample wins against another sample, the score is considered to be
1.
Otherwise the score is 0.  Then the average score would be rate of
winning.
Deviation of a won comparison would be (1 - score).  It is (0 - score)
in lost case.
Thus, variance of scores for a sample would be
((1 - score)^2 * numWins + (0 - score)^2 * numLosts) / numComparisons